Rolled SuperMatter.dm back to an old version.
The "new" version you guys were working on is included in full in
a giant comment at the bottom of the file.  Just putting this back
on so the game is playable until you guys can get the engine back
to a point where it's not embarassing to have on the server.